+++ severe/src/roblox/data/vector3.lua
